Fix missing smells when ignore rules are used in qlty.toml
The following qlty.toml config will cause IgnoreGroup to register a blanket wildcard (`*`) ignore:

```
[[ignore]]
rules = ["tool:rule_name"]
```

Due to the way the qlty.toml is loaded, the `file_patterns` for this ignore definition will be set to `*`, thereby creating an ignore group that looks like:

```
["**/protos/**", "**/other_ignore/**", "*", "**/tailwind.css", ...]
```

Which will match (and ignore) all files.

The proposed fix is to skip an ignore wildcard if any rules are defined, as this should not be treated as a blanket workspace_entry ignore (an invocation must be run to determine what needs to be ignored).

There is a possible optimization for `tool:*` rules for the plugin workspace entry matcher, but I did not want to complicate the implementation.