Fix bubbling of stderr to tool install log files
Running duct::cmd::run() by default bails on non-zero exit codes, avoiding capture of any stderr. This change updates to run with `unchecked()` so that stderr can be captured into log files.

Due to this change, the specifics of Tool.run_command had to be updated to extract the original command for error display to maintain the original (somewhat helpful errors), as the Err object containing the command must now be synthesized manually (duct returned the err in the form `["program", "arg1", ...]` but this list of program/args is not available on the Expression object except via before_spawn).