ruby(macos): add preflight dependency checks to Ruby binary install
We need gmp, libyaml, and openssl in order to install Ruby on any macOS system. This pre-flight check (ported from #2007) validates that the libraries are installed via Homebrew.

This PR also adds this check for x86_64 architectures, as well as disables retries for the Ruby tool installer. If we don't disable retries, this (slow) check will be run 3+ times (9 total brew checks).

Note that an improvement of this approach would be to simply request permission from the user to install these libraries instead of exiting. This could be implemented as a follow-up.