Add `biome.jsonc` as a default config file for Biome
Biome allows for its config file to be either a normal JSON file _or_ JSON with comments (which uses the `.jsonc` file extension). Support for JSONC is prominently mentioned in [Biome's Getting Started guide](https://biomejs.dev/guides/getting-started/#configuration), so it's reasonable to assume that a non-trivial number of people have opted to use it. This plugin should offer it as one of the default config files.