Make qlty install more context aware and inline with check (#2064)
There were more than a few things different between the tool
installation process in qlty install (which was more barebones) and qlty
check which was more comprehensive.

One of which was managing package_files and filters properly causing
qlty install to not work as intended in many instances.

This PR aims to make them behave similar as possible while minimizing
avoidable duplication.