feat: Add support for `--sarif` flag to `qlty smells` command (#2037)
This PR adds support for generating SARIF output using the --sarif flag
when running the `qlty smells` command.


---

Notes:

I'm not an expert in Rust, and I encountered a few challenges while
implementing this feature, so happy for any comments to improve it.

The main issue was that the existing `SarifFormatter` only supported a
single type: `qlty_check::Report`. Since Rust doesn't have inheritance,
I resolved this by extracting a shared trait (`SarifTrait`) and
implementing it for both `qlty_check::Report` and
`qlty_analysis::Report`. The `SarifFormatter` was then updated to accept
any type that implements this trait.

I'm not completely confident that all the structs and traits are placed
in the most appropriate modules, so I'm open to feedback or suggestions
about how to better organize them according to project conventions or
best practices.