chore: regenerate protobufs for behavior neutral diff (#2259)
I'm adding a new field shortly, but noticed, in generating, that there
was a lot of added noise in the diff. We must have updated a tool
version, since now re-generating protobufs with the same definition
yields this diff. By first committing and merging we can have a cleaner
diff for actual, intended changes.