feat: Support merge group detection on any CI provider listening to GitHub push events (#2320)
In #2296 we added detection of merge groups. When `qlty coverage
publish` is run as part of a merge group build, it sends back metadata
to Qlty informing it that it's "reference type" is a merge group. This
allows us to post back commit statuses for merge groups commits,
something we previously only did for pull requests commits.

However, the detection algorithm -- checking for the presence of an
environment variable -- was specific to GitHub. This PR changes the
implementation to instead check if the branch name matches the specific
convention GitHub uses for merge group branch names (If a PR is added to
a merge queue, GItHub automatically creates a branch named matching the
pattern `gh-readonly-queue/*`). While perhaps more fragile (anyone could
name their branches this way and it would get erroneously detected as
being part of a merge group), this GitHub's recommended way of
[integrating with 3rd party CI
vendors](https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-a-merge-queue#triggering-merge-group-checks-with-third-party-ci-providers).