fix: Remove `./` from beginning of `--add-prefix` (#2376)
Addresses primary bug identified in #2375 while leaving some of what is
discussed not yet addressed.

Also adds a couple tests to document current behavior that I found a bit
surprising: notably that `--add-prefix` treats its argument more like a
string (`foo` and `foo/` are different) and `--strip-prefix` is more
path-aware and treats `foo` and `foo/` as the same path component.