Add support for Jacoco reports with group elements (#2572)
Jacoco XML supports an optional "group" element that wraps packages. I
believe this is referred to as "multi-module" support, and it can be
turned on by using the "report-aggregate goal in a Maven".

The lack of support for groups goes back apparently to the old test
reporter, which also did
[not](https://github.com/qltysh-archive/test-reporter/issues/376)
[support](https://github.com/qltysh-archive/test-reporter/issues/409)
[it](https://github.com/qltysh-archive/test-reporter/issues/434).

This adds support for jacoco.xml files that have group elements.

Note that a group contains a name attribute. I'm not certain but I
believe the name may map to a folder, which would obviate the need for
to specify a JACOCO_SOURCE_PATH for each group. This PR does not
implement that behavior but it's something we could consider for the
future.