fix: Use single quotes to prevent $ escape in windows command (#2589)
https://github.com/qltysh/qlty/actions/runs/20342268234/job/58446611866

> = : The term '=' is not recognized as the name of a cmdlet, function,
script file, or operable program. Check the
spelling of the name, or if a path was included, verify that the path is
correct and try again.
At line:1 char:2
> +  = 'SilentlyContinue'; iwr https://qlty.sh/ | iex
> +  ~

Suggests that `$` character is being interpreted by the shell (sh)
before reaching PowerShell.
Switching to using single quotes to prevent that.