fix: Use mypy json output (#2609)
As of mypy v 1.11 they have supported a json output format.

So instead of using regex which can have issues I believe we should
switch to using json output which is more consistent.

The column in json are zero indexed instead of one though, and can be -1
in case of missing column info.

Also allows us to bump known good version.